refactor: split options.tsx into containers and templates
